requirements.txt
requests
Pillow
gradio
modules